pyproject.toml
[project]
name = "rsi-trading-agent-v2"
version = "0.1.0"
description = "Week 1 BTC model-based research prototype with held-out evaluation"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "langgraph>=0.6.0",
  "polars>=1.0.0",
  "ta-lib>=0.4.32",
  "vectorbt>=0.27.0",
  "optuna>=4.0.0",
  "ccxt>=4.4.0",
  "pyarrow>=15.0.0",
  "pyyaml>=6.0.1",
  "shap>=0.45.0",
  "lightgbm>=4.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = false

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]
